notification testing and refactor
